refactor ready.js to resolve circular dependency
